+++ src/index.ts
@@ -5,6 +5,7 @@ export { default as PasswordRecoveryForm} from './components/login/PasswordRecov
 
 export { default as RecaptchaForm} from './components/recaptcha/RecaptchaForm';
 
+export { default as NotificationSnackbar } from './components/NotificationSnackbar';
 export { default as TextFieldValidate } from './components/TextFieldValidate';
 
 export type { BorderProps } from "./types/BorderProps";