+++ CHANGELOG.md
@@ -6,5 +6,6 @@ e este projeto adota [Versionamento Semântico](https://semver.org/lang/pt-BR/).
 
 ## [1.0.0] - 2025-12-28
 - Criação do package
+- Implementacao do componente NotificationSnackbar.